backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
anthropic>=0.40.0
python-multipart>=0.0.6
slowapi>=0.1.9
python-dotenv>=1.0.0
